requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
idna==2.10
PySocks==1.7.1
PyYAML==5.3.1
requests==2.24.0
shodan==1.23.0
urllib3==1.25.10
XlsxWriter==1.3.2
